src folder kedro
This the src folder used as exploration to use kedro for a load_data pipeline. 
Not the full kedro folder could be added because the datasets are too large to add to github. 
However, a folder "data" was established in the root folder of this folder, containing the initial datasets and the output file from our LLM. Together with a cleaned_df and a full_df. In the main branch you find a further worked out kedro setup for both the advanced random forest and xgboost models.